A2A Server Weather Agent with Google Cloud Run/pyproject.toml

[project]
name = "a2a-weather-agent"
version = "0.1.0"
description = "A2A Weather Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain>=0.2.0",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "langchain-openai>=0.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
    "a2a-sdk>=0.3.0",
    "fastmcp>=0.1.0",
    "nest-asyncio>=1.6.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


